// src/models/device.ts
/**
 * Copyright 2025. ThingsBoard
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *  http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { CustomerId, DeviceId, DeviceProfileId, JsonNode, OtaPackageId, TenantId } from '../types/common';

export interface DeviceData {
  configuration?: JsonNode;
  transportConfiguration?: JsonNode;
}

export interface Device {
  id?: DeviceId;
  createdTime?: number;
  tenantId?: TenantId;
  customerId?: CustomerId;
  name: string;
  type: string;
  label?: string;
  deviceProfileId: DeviceProfileId;
  firmwareId?: OtaPackageId;
  softwareId?: OtaPackageId;
  version?: number;
  additionalInfo?: JsonNode;
  deviceData?: DeviceData;
}

export interface DeviceInfo extends Device {
  customerTitle?: string;
  customerIsPublic?: boolean;
  deviceProfileName?: string;
}

export interface PageDataDevice {
  data: Device[];
  totalPages: number;
  totalElements: number;
  hasNext: boolean;
}